orchestrator/requirements.txt
anthropic>=0.25.0
requests>=2.31.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
